fix(in-app-agent): wrap MCP tool results as LanguageModelV3 json
Providers serialize output.value. Unwrapping MCP content and listing
known AI SDK types is extra; wrapping the execute result is enough.
