+++ packages/shared/src/in-app-agent/server/toolResults.ts
@@ -72,26 +72,15 @@ export function isSilentInAppAgentMcpToolOutput(
   );
 }
 
-const AI_SDK_TOOL_MODEL_OUTPUT_TYPES = [
-  "text",
-  "json",
-  "content",
-  "error-text",
-  "error-json",
-] as const;
-
-type AiSdkToolModelOutputType = (typeof AI_SDK_TOOL_MODEL_OUTPUT_TYPES)[number];
-
 export type AiSdkToolModelOutput = {
-  type: AiSdkToolModelOutputType;
+  type: string;
   value: unknown;
 };
 
 /**
- * Map MCP CallToolResult envelopes (and silent pointers) to the AI SDK
- * `{ type, value }` tool-result shape. Bedrock Converse only serializes
- * `output.value`; a raw `{ content: [...] }` envelope becomes an empty
- * `toolResult` on the wire.
+ * Map tool execute results to the LanguageModelV3 `{ type, value }` shape.
+ * Providers serialize `output.value`; a raw MCP `{ content }` envelope has
+ * no `.value` and becomes an empty tool result on the wire.
  */
 export function toAiSdkToolModelOutput(output: unknown): AiSdkToolModelOutput {
   if (typeof output === "string") {
@@ -112,44 +101,15 @@ export function toAiSdkToolModelOutput(output: unknown): AiSdkToolModelOutput {
     return toAiSdkToolModelOutput(output.output);
   }
 
-  if (isAiSdkToolModelOutput(output)) {
-    return output;
-  }
-
-  const mcpText = getMcpContentText(output);
-  if (mcpText !== undefined) {
-    const isError =
-      isRecord(output) && (output.isError === true || output.error === true);
-
-    return { type: isError ? "error-text" : "text", value: mcpText };
+  if (
+    isRecord(output) &&
+    typeof output.type === "string" &&
+    "value" in output
+  ) {
+    return { type: output.type, value: output.value };
   }
 
   return { type: "json", value: output };
-}
-
-function isAiSdkToolModelOutput(value: unknown): value is AiSdkToolModelOutput {
-  return (
-    isRecord(value) &&
-    typeof value.type === "string" &&
-    AI_SDK_TOOL_MODEL_OUTPUT_TYPES.includes(
-      value.type as AiSdkToolModelOutputType,
-    ) &&
-    "value" in value
-  );
-}
-
-function getMcpContentText(output: unknown): string | undefined {
-  if (!isRecord(output) || !Array.isArray(output.content)) {
-    return undefined;
-  }
-
-  const texts = output.content.flatMap((part) =>
-    isRecord(part) && part.type === "text" && typeof part.text === "string"
-      ? [part.text]
-      : [],
-  );
-
-  return texts.length > 0 ? texts.join("\n") : undefined;
 }
 
 /** Withhold private persisted event payloads from browser-facing streams. */
